PHPStan: Add phpstan support

+++ tests/bootstrap.php
@@ -14,7 +14,7 @@ namespace Gamajo\PluginSlug\Tests;
 
 // Check for a `--testsuite integration` arg when calling phpunit, and use it to conditionally load up WordPress.
 $plugin_slug_argv = $GLOBALS['argv'];
-$plugin_slug_key  = array_search( '--testsuite', $plugin_slug_argv, true );
+$plugin_slug_key  = (int) array_search( '--testsuite', $plugin_slug_argv, true );
 
 if ( $plugin_slug_key && 'integration' === $plugin_slug_argv[ $plugin_slug_key + 1 ] ) {
 	$plugin_slug_tests_dir = getenv( 'WP_TESTS_DIR' );
